Fix signed/unsigned warnings

+++ src/operator/elemwise_op_common.h
@@ -74,8 +74,8 @@ template<int n_in, int n_out>
 inline bool ElemwiseShape(const nnvm::NodeAttrs& attrs,
                           std::vector<TShape> *in_attrs,
                           std::vector<TShape> *out_attrs) {
-  CHECK_EQ(in_attrs->size(), n_in) << " in operator " << attrs.name;
-  CHECK_EQ(out_attrs->size(), n_out);
+  CHECK_EQ(in_attrs->size(), (size_t)n_in) << " in operator " << attrs.name;
+  CHECK_EQ(out_attrs->size(), (size_t)n_out);
   return ElemwiseAttr<TShape, shape_is_none, true>(
     attrs, in_attrs, out_attrs);
 }
@@ -88,8 +88,8 @@ template<int n_in, int n_out>
 inline bool ElemwiseType(const nnvm::NodeAttrs& attrs,
                          std::vector<int> *in_attrs,
                          std::vector<int> *out_attrs) {
-  CHECK_EQ(in_attrs->size(), n_in) << " in operator " << attrs.name;
-  CHECK_EQ(out_attrs->size(), n_out);
+  CHECK_EQ(in_attrs->size(), (size_t)n_in) << " in operator " << attrs.name;
+  CHECK_EQ(out_attrs->size(), (size_t)n_out);
   return ElemwiseAttr<int, type_is_none, true>(
     attrs, in_attrs, out_attrs);
 }
